Fix type errors and simplified some code
Most of the changes were minor type changes to make sure the right type
was being converted and returned.
I also simplified some of the lines, where there were some redudencies.
Some notes, unless the comment is telling you something important, it
really should not be included in the merge to master. All of the comment
out lines for test data should be removed before merging.