Change list names
I changed the lists names so they were more conventional, uppercase
first letter should only be used by classes, and lists should be plural.
I also had to change the ó to a o because there were some problems
displaying it on console (It might actually be fine on Hive so maybe in
the future we can change it back).