Clean up code
Changed payment and web to more descriptive names to better convey what
they are. Changed method names to not start with uppercase, the
convention is to only capitalize class names. I also fixed a bug with
products_ordered where there was a possibility for no products to be
generated. I also explicity declared types in some places, this was for
my own sanity while working on EcommerceGenerator.scala.