Fix null rows bug and compatibility issues
When uploading to hive there should not longer be any problems where
there are null rows. This was fixed, I am pretty sure, by putting the
csv on hdfs in its own directory, so other files in the same directory
don't get leached in. Also fixed the defualt database name.